Reviewed for publishing
Included the title to reflect the actual name of the document, expanded
introduction, worked on the flow of the document and removed rawgit
reference. Other changes were made due to grammar, structural and
coding errors. Created an acknowledgments section to include a ‘thank
you message’ for a contributor, which was listed under references.
Also, the original author would need to correctly cite references
within the body as there is no indication where they belong.